Chore: Move the implementation to cpp
